"tvb_reported_length_remaining()" returns a "gint", so assign its return value to a "gint".
svn path=/trunk/; revision=12223

+++ epan/dissectors/packet-icep.c
@@ -544,7 +544,7 @@ static void dissect_ice_params(proto_tree *tree, tvbuff_t *tvb,
 	 */
 	
 	gint32 size = 0;
-	guint32 tvb_data_remained = 0;
+	gint tvb_data_remained = 0;
 	
 	(*consumed) = 0;
 	
@@ -609,7 +609,7 @@ static void dissect_ice_params(proto_tree *tree, tvbuff_t *tvb,
 	/* check if I got all encapsulated data */
 	tvb_data_remained = tvb_reported_length_remaining(tvb, offset);
 	
-	if ( ((gint32) tvb_data_remained) < ( size - ICEP_MIN_PARAMS_SIZE ) ) {
+	if ( tvb_data_remained < ( size - ICEP_MIN_PARAMS_SIZE ) ) {
 		
 		if (tree)
 			proto_tree_add_text(tree, tvb, offset, -1,
